+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/service/AqlQueryServiceImp.java
@@ -17,12 +17,16 @@
  */
 package org.ehrbase.openehr.aqlengine.service;
 
+import static org.ehrbase.openehr.aqlengine.AqlParameterReplacement.replaceParameters;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.re2j.Pattern;
 import java.lang.constant.Constable;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -31,14 +35,14 @@ import java.util.Optional;
 import java.util.stream.LongStream;
 import java.util.stream.Stream;
 import org.apache.commons.collections4.CollectionUtils;
+import org.ehrbase.api.dto.AqlQueryContext;
+import org.ehrbase.api.dto.AqlQueryRequest;
 import org.ehrbase.api.exception.AqlFeatureNotImplementedException;
 import org.ehrbase.api.exception.BadGatewayException;
 import org.ehrbase.api.exception.IllegalAqlException;
 import org.ehrbase.api.exception.InternalServerException;
-import org.ehrbase.api.exception.InvalidApiParameterException;
-import org.ehrbase.api.service.AqlQueryRequest;
+import org.ehrbase.api.exception.UnprocessableEntityException;
 import org.ehrbase.api.service.AqlQueryService;
-import org.ehrbase.openehr.aqlengine.AqlParameterReplacement;
 import org.ehrbase.openehr.aqlengine.AqlQueryUtils;
 import org.ehrbase.openehr.aqlengine.asl.AqlSqlLayer;
 import org.ehrbase.openehr.aqlengine.asl.model.query.AslRootQuery;
@@ -47,6 +51,7 @@ import org.ehrbase.openehr.aqlengine.querywrapper.AqlQueryWrapper;
 import org.ehrbase.openehr.aqlengine.querywrapper.select.SelectWrapper;
 import org.ehrbase.openehr.aqlengine.querywrapper.select.SelectWrapper.SelectType;
 import org.ehrbase.openehr.aqlengine.repository.AqlQueryRepository;
+import org.ehrbase.openehr.aqlengine.repository.PreparedQuery;
 import org.ehrbase.openehr.sdk.aql.dto.AqlQuery;
 import org.ehrbase.openehr.sdk.aql.dto.containment.AbstractContainmentExpression;
 import org.ehrbase.openehr.sdk.aql.dto.containment.ContainmentClassExpression;
@@ -78,17 +83,23 @@ public class AqlQueryServiceImp implements AqlQueryService {
     private final ExternalTerminologyValidation tsAdapter;
     private final AqlSqlLayer aqlSqlLayer;
     private final AqlQueryFeatureCheck aqlQueryFeatureCheck;
+    private final ObjectMapper objectMapper;
+    private final AqlQueryContext aqlQueryContext;
 
     @Autowired
     public AqlQueryServiceImp(
             AqlQueryRepository aqlQueryRepository,
             ExternalTerminologyValidation tsAdapter,
             AqlSqlLayer aqlSqlLayer,
-            AqlQueryFeatureCheck aqlQueryFeatureCheck) {
+            AqlQueryFeatureCheck aqlQueryFeatureCheck,
+            ObjectMapper objectMapper,
+            AqlQueryContext aqlQueryContext) {
         this.aqlQueryRepository = aqlQueryRepository;
         this.tsAdapter = tsAdapter;
         this.aqlSqlLayer = aqlSqlLayer;
         this.aqlQueryFeatureCheck = aqlQueryFeatureCheck;
+        this.objectMapper = objectMapper;
+        this.aqlQueryContext = aqlQueryContext;
     }
 
     @Override
@@ -99,103 +110,140 @@ public class AqlQueryServiceImp implements AqlQueryService {
     private QueryResultDto queryAql(AqlQueryRequest aqlQueryRequest) {
         // TODO: check that select aliases are not duplicated
         try {
-            AqlQuery aqlQuery = AqlQueryParser.parse(aqlQueryRequest.queryString());
-            // apply limit and offset - where the definitions from the aql will be preferred over the request values.
-            aqlQuery.setLimit(Optional.ofNullable(aqlQuery.getLimit()).orElse(aqlQueryRequest.fetch()));
-            aqlQuery.setOffset(Optional.ofNullable(aqlQuery.getOffset()).orElse(aqlQueryRequest.offset()));
-
-            // Sanity check - In AQL there is no offset without limit.
-            if (aqlQuery.getOffset() != null && aqlQuery.getLimit() == null) {
-                throw new InvalidApiParameterException(
-                        "Invalid AQL query: provided offset %s without a limit".formatted(aqlQuery.getOffset()));
-            }
-
-            AqlParameterReplacement.replaceParameters(aqlQuery, aqlQueryRequest.parameters());
-
-            replaceEhrPaths(aqlQuery);
+            AqlQuery aqlQuery = buildAqlQuery(aqlQueryRequest);
 
             aqlQueryFeatureCheck.ensureQuerySupported(aqlQuery);
 
             try {
+                if (logger.isTraceEnabled()) {
+                    logger.trace(objectMapper.writeValueAsString(aqlQuery));
+                }
+
                 AqlQueryWrapper queryWrapper = AqlQueryWrapper.create(aqlQuery);
 
                 AslRootQuery aslQuery = aqlSqlLayer.buildAslRootQuery(queryWrapper);
                 List<SelectWrapper> nonPrimitiveSelects =
                         queryWrapper.nonPrimitiveSelects().toList();
-                List<List<Object>> result = aqlQueryRepository.executeQuery(aslQuery, nonPrimitiveSelects);
 
-                if (nonPrimitiveSelects.isEmpty()) {
-                    // only primitives selected: only a count() was performed, so the list must be constructed
-                    result = LongStream.range(0, (long) result.getFirst().getFirst())
-                            .<List<Object>>mapToObj(i -> new ArrayList<>())
-                            .toList();
+                PreparedQuery preparedQuery = aqlQueryRepository.prepareQuery(aslQuery, nonPrimitiveSelects);
+
+                // aql debug options
+                if (aqlQueryContext.showExecutedSql()) {
+                    aqlQueryContext.setMetaProperty(
+                            AqlQueryContext.EhrbaseMetaProperty.EXECUTED_SQL,
+                            AqlQueryRepository.getQuerySql(preparedQuery));
+                }
+                if (aqlQueryContext.showQueryPlan()) {
+                    // for dry-run omit analyze
+                    boolean analyze = !aqlQueryContext.isDryRun();
+                    String explainedQuery = aqlQueryRepository.explainQuery(analyze, preparedQuery);
+                    TypeReference<HashMap<String, Object>> typeRef = new TypeReference<>() {};
+                    aqlQueryContext.setMetaProperty(
+                            AqlQueryContext.EhrbaseMetaProperty.QUERY_PLAN,
+                            objectMapper.readValue(explainedQuery, typeRef));
                 }
 
-                List<SelectWrapper> selects = queryWrapper.selects();
-                // Since we do not add primitive value selects to the SQL query, we add them after the query was
-                // executed
-                for (int i = 0; i < selects.size(); i++) {
-                    SelectWrapper sd = selects.get(i);
-                    if (sd.type() == SelectType.PRIMITIVE) {
-                        Constable value = sd.getPrimitive().getValue();
-                        for (List<Object> row : result) {
-                            row.add(i, value);
-                        }
-                    }
+                if (aqlQueryContext.showExecutedAql()) {
+                    aqlQueryContext.setExecutedAql(AqlRenderer.render(aqlQuery));
                 }
 
-                if (logger.isTraceEnabled()) {
-                    try {
-                        logger.trace(new ObjectMapper().writeValueAsString(aqlQuery));
-                    } catch (JsonProcessingException e) {
-                        throw new RuntimeException(e.getMessage(), e);
-                    }
+                Optional.of(queryWrapper)
+                        .map(AqlQueryWrapper::limit)
+                        .map(Long::intValue)
+                        .ifPresent(limit -> {
+                            aqlQueryContext.setMetaProperty(AqlQueryContext.EhrbaseMetaProperty.FETCH, limit);
+                            // in case only a limit was used we define the default offset as 0
+                            aqlQueryContext.setMetaProperty(
+                                    AqlQueryContext.EhrbaseMetaProperty.OFFSET,
+                                    Optional.of(queryWrapper)
+                                            .map(AqlQueryWrapper::offset)
+                                            .map(Long::intValue)
+                                            .orElse(0));
+                        });
+
+                List<List<Object>> resultData;
+                if (aqlQueryContext.isDryRun()) {
+                    resultData = List.of();
+                } else {
+                    resultData = executeQuery(preparedQuery, queryWrapper, nonPrimitiveSelects);
+                    aqlQueryContext.setMetaProperty(AqlQueryContext.EhrbaseMetaProperty.RESULT_SIZE, resultData.size());
                 }
+                return formatResult(queryWrapper.selects(), resultData);
 
-                String understoodByAqlParser = AqlRenderer.render(aqlQuery);
-                return formatResult(queryWrapper, result, understoodByAqlParser);
-
-            } catch (IllegalArgumentException e) {
+            } catch (IllegalArgumentException | JsonProcessingException e) {
                 // regular IllegalArgumentException, not due to illegal query parameters
                 throw new InternalServerException(e.getMessage(), e);
             }
-
         } catch (RestClientException e) {
-            throw new BadGatewayException(
-                    "Bad gateway: %s"
-                            .formatted(Optional.of(e)
-                                    .map(Throwable::getCause)
-                                    .orElse(e)
-                                    .getMessage()),
-                    e);
+            throw new BadGatewayException(errorMessage("Bad gateway", e), e);
         } catch (DataAccessException e) {
-            throw new InternalServerException(
-                    "Data Access Error: %s"
-                            .formatted(Optional.of(e)
-                                    .map(Throwable::getCause)
-                                    .orElse(e)
-                                    .getMessage()),
-                    e);
+            throw new InternalServerException(errorMessage("Data Access Error", e), e);
         } catch (AqlParseException e) {
-            throw new IllegalAqlException(
-                    "Could not parse AQL query: %s"
-                            .formatted(Optional.of(e)
-                                    .map(Throwable::getCause)
-                                    .orElse(e)
-                                    .getMessage()),
-                    e);
+            throw new IllegalAqlException(errorMessage("Could not parse AQL query", e), e);
         }
     }
 
-    private QueryResultDto formatResult(AqlQueryWrapper query, List<List<Object>> resultData, String queryString) {
+    static AqlQuery buildAqlQuery(AqlQueryRequest aqlQueryRequest) {
 
-        List<SelectWrapper> selectFields = query.selects();
+        AqlQuery aqlQuery = AqlQueryParser.parse(aqlQueryRequest.queryString());
 
-        QueryResultDto dto = new QueryResultDto();
-        dto.setExecutedAQL(queryString);
+        // apply limit and offset - where the definitions from the aql are the precedence
+        Long fetchParam = aqlQueryRequest.fetch();
+        Long offsetParam = aqlQueryRequest.offset();
+        Long limitQuery = aqlQuery.getLimit();
 
-        Optional.ofNullable(query.limit()).ifPresent(v -> dto.setLimit(v.intValue()));
-        Optional.ofNullable(query.offset()).ifPresent(v -> dto.setOffset(v.intValue()));
+        // verify not parameter fetch offset are defined when query contains a LIMIT or assign fetch parameter
+        if (limitQuery == null) {
+            aqlQuery.setLimit(fetchParam);
+            aqlQuery.setOffset(offsetParam);
+        } else {
+            if (fetchParam != null || offsetParam != null) {
+                throw new UnprocessableEntityException(
+                        "Query contains a LIMIT clause, fetch and offset parameters must not be used");
+            }
+        }
+
+        // sanity check parameter
+        if (aqlQuery.getOffset() != null && aqlQuery.getLimit() == null) {
+            throw new UnprocessableEntityException(
+                    "Query parameter for offset %s provided without a fetch limit".formatted(aqlQuery.getOffset()));
+        }
+
+        // postprocess
+        replaceParameters(aqlQuery, aqlQueryRequest.parameters());
+        replaceEhrPaths(aqlQuery);
+
+        return aqlQuery;
+    }
+
+    private List<List<Object>> executeQuery(
+            PreparedQuery preparedQuery, AqlQueryWrapper queryWrapper, List<SelectWrapper> nonPrimitiveSelects) {
+
+        List<List<Object>> resultData = aqlQueryRepository.executeQuery(preparedQuery);
+
+        if (nonPrimitiveSelects.isEmpty()) {
+            // only primitives selected: only a count() was performed, so the list must be constructed
+            resultData = LongStream.range(0, (long) resultData.getFirst().getFirst())
+                    .<List<Object>>mapToObj(i -> new ArrayList<>())
+                    .toList();
+        }
+
+        List<SelectWrapper> selects = queryWrapper.selects();
+        // Since we do not add primitive value selects to the SQL query, we add them after the query was
+        // executed
+        for (int i = 0; i < selects.size(); i++) {
+            SelectWrapper sd = selects.get(i);
+            if (sd.type() == SelectType.PRIMITIVE) {
+                Constable value = sd.getPrimitive().getValue();
+                for (List<Object> row : resultData) {
+                    row.add(i, value);
+                }
+            }
+        }
+        return resultData;
+    }
+
+    private QueryResultDto formatResult(List<SelectWrapper> selectFields, List<List<Object>> resultData) {
 
         Map<String, String> columns = new LinkedHashMap<>();
         for (int i = 0; i < selectFields.size(); i++) {
@@ -204,6 +252,8 @@ public class AqlQueryServiceImp implements AqlQueryService {
                     Optional.of(namePath).map(SelectWrapper::getSelectAlias).orElse("#" + i),
                     namePath.getSelectPath().orElse(null));
         }
+
+        QueryResultDto dto = new QueryResultDto();
         dto.setVariables(columns);
 
         List<ResultHolder> resultList = resultData.stream()
@@ -229,8 +279,6 @@ public class AqlQueryServiceImp implements AqlQueryService {
      * Rephrases EHR.composition and EHR.status CONTAINS statements so that they can be handled regularly by the aql engine.
      * E.g. <code>SELECT e/ehr_status FROM EHR</code> is rewritten as <code>SELECT s FROM EHR e CONTAINS EHR_STATUS s</code>,
      * <code>SELECT e/composition FROM EHR</code> is rewritten as <code>SELECT c FROM EHR e CONTAINS COMPOSITION c</code>.
-     *
-     * @param aqlQuery
      */
     static void replaceEhrPaths(AqlQuery aqlQuery) {
         replaceEhrPath(aqlQuery, "compositions", "COMPOSITION", "c");
@@ -240,8 +288,6 @@ public class AqlQueryServiceImp implements AqlQueryService {
     /**
      * Rephrases a path from EHR to EHR_STATUS as CONTAINS statement so that it can be handled regularly by the aql engine.
      * E.g. <code>SELECT e/status FROM EHR</code> is rewritten as <code>SELECT s FROM EHR e CONTAINS EHR_STATUS s</code>.
-     *
-     * @param aqlQuery
      */
     static void replaceEhrPath(AqlQuery aqlQuery, String ehrPath, String type, String aliasPrefix) {
 
@@ -334,5 +380,9 @@ public class AqlQueryServiceImp implements AqlQueryService {
             List<PathNode> pathNodes = ip.getPath().getPathNodes();
             ip.setPath(pathNodes.size() == 1 ? null : new AqlObjectPath(pathNodes.subList(1, pathNodes.size())));
         });
+    }
+
+    private static String errorMessage(String prefix, Exception e) {
+        return prefix + ": " + Optional.of(e).map(Throwable::getCause).orElse(e).getMessage();
     }
 }