+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/OpenehrQueryController.java
@@ -28,23 +28,20 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Stream;
+import org.ehrbase.api.dto.AqlQueryContext;
+import org.ehrbase.api.dto.AqlQueryRequest;
 import org.ehrbase.api.exception.InvalidApiParameterException;
 import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.api.rest.HttpRestContext;
-import org.ehrbase.api.service.AqlQueryRequest;
 import org.ehrbase.api.service.AqlQueryService;
-import org.ehrbase.api.service.StatusService;
 import org.ehrbase.api.service.StoredQueryService;
-import org.ehrbase.openehr.sdk.response.dto.MetaData;
 import org.ehrbase.openehr.sdk.response.dto.QueryResponseData;
 import org.ehrbase.openehr.sdk.response.dto.ehrscape.QueryDefinitionResultDto;
 import org.ehrbase.openehr.sdk.response.dto.ehrscape.QueryResultDto;
-import org.ehrbase.openehr.sdk.util.rmconstants.RmConstants;
 import org.ehrbase.rest.BaseController;
 import org.ehrbase.rest.openehr.specification.QueryApiSpecification;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -69,24 +66,23 @@ import org.springframework.web.bind.annotation.RestController;
         produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
 public class OpenehrQueryController extends BaseController implements QueryApiSpecification {
 
+    // request parameter
     private static final String QUERY_PARAMETERS = "query_parameters";
-    private static final String FETCH = "fetch";
-    private static final String OFFSET = "offset";
+    private static final String FETCH_PARAM = "fetch";
+    private static final String OFFSET_PARAM = "offset";
+    private static final String Q_PARAM = "q";
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
     private final AqlQueryService aqlQueryService;
     private final StoredQueryService storedQueryService;
-    private final StatusService statusService;
-
-    @Value("${ehrbase.rest.aql.response.generator-details-enabled:false}")
-    boolean generatorDetailsEnabled = false;
+    private final AqlQueryContext aqlQueryContext;
 
     public OpenehrQueryController(
-            AqlQueryService aqlQueryService, StoredQueryService storedQueryService, StatusService statusService) {
+            AqlQueryService aqlQueryService, StoredQueryService storedQueryService, AqlQueryContext aqlQueryContext) {
         this.aqlQueryService = aqlQueryService;
         this.storedQueryService = storedQueryService;
-        this.statusService = statusService;
+        this.aqlQueryContext = aqlQueryContext;
     }
 
     /**
@@ -95,10 +91,9 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
     @Override
     @GetMapping(path = "/aql")
     public ResponseEntity<QueryResponseData> executeAdHocQuery(
-            @RequestParam(name = "q") String queryString,
-            // FIXME possible validation for [offset|row] >= 0
-            @RequestParam(name = OFFSET, required = false) Integer offset,
-            @RequestParam(name = FETCH, required = false) Integer fetch,
+            @RequestParam(name = Q_PARAM) String queryString,
+            @RequestParam(name = OFFSET_PARAM, required = false) Integer offset,
+            @RequestParam(name = FETCH_PARAM, required = false) Integer fetch,
             @RequestParam(name = QUERY_PARAMETERS, required = false) Map<String, Object> queryParameters,
             @RequestHeader(name = ACCEPT, required = false) String accept) {
 
@@ -106,15 +101,16 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
         HttpRestContext.register(QUERY_EXECUTE_ENDPOINT, Boolean.TRUE);
 
         // get the query and pass it to the service
-        QueryResultDto queryResult = invoke(
+        AqlQueryRequest aqlQueryRequest = createRequest(
                 queryString,
                 queryParameters,
                 Optional.ofNullable(fetch).map(Integer::longValue),
                 Optional.ofNullable(offset).map(Integer::longValue));
+        QueryResultDto aqlQueryResult = aqlQueryService.query(aqlQueryRequest);
 
         // create and return response
         QueryResponseData queryResponseData =
-                createQueryResponse(queryString, queryResult, createLocationUri("query", "aql"));
+                createQueryResponse(aqlQueryResult, queryString, createLocationUri("query", "aql"));
 
         return ResponseEntity.ok(queryResponseData);
     }
@@ -134,23 +130,24 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
         logger.debug("Got following input: {}", queryRequest);
 
         // sanity check
-        Object rawQuery = queryRequest.get("q");
-        if (rawQuery == null) {
-            throw new InvalidApiParameterException("No aql query provided");
-        }
-        if (rawQuery instanceof ArrayList<?>) {
-            throw new InvalidApiParameterException("Multiple aql queries provided");
-        }
-        String queryString = (String) rawQuery;
+        Object rawQuery = queryRequest.get(Q_PARAM);
+        String queryString =
+                switch (rawQuery) {
+                    case null -> throw new InvalidApiParameterException("No aql query provided");
+                    case ArrayList<?> __ -> throw new InvalidApiParameterException("Multiple aql queries provided");
+                    case String s -> s;
+                    default -> throw new InvalidApiParameterException("Data type of aql query not supported");
+                };
 
         // Enriches request attributes with aql for later audit processing
         HttpRestContext.register(QUERY_EXECUTE_ENDPOINT, Boolean.TRUE);
 
         // get the query and pass it to the service
-        QueryResultDto queryResult = invoke(queryString, queryRequest);
+        AqlQueryRequest aqlQueryRequest = createRequest(queryString, queryRequest);
+        QueryResultDto aqlQueryResult = aqlQueryService.query(aqlQueryRequest);
 
         // create and return response
-        QueryResponseData queryResponseData = createQueryResponse(queryString, queryResult);
+        QueryResponseData queryResponseData = createQueryResponse(aqlQueryResult, queryString, null);
 
         return ResponseEntity.ok(queryResponseData);
     }
@@ -163,8 +160,8 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
     public ResponseEntity<QueryResponseData> executeStoredQuery(
             @PathVariable(name = "qualified_query_name") String qualifiedQueryName,
             @PathVariable(name = "version", required = false) String version,
-            @RequestParam(name = OFFSET, required = false) Integer offset,
-            @RequestParam(name = FETCH, required = false) Integer fetch,
+            @RequestParam(name = OFFSET_PARAM, required = false) Integer offset,
+            @RequestParam(name = FETCH_PARAM, required = false) Integer fetch,
             @RequestParam(name = QUERY_PARAMETERS, required = false) Map<String, Object> queryParameters,
             @RequestHeader(name = ACCEPT, required = false) String accept) {
 
@@ -184,11 +181,12 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
         String queryString = queryDefinition.getQueryText();
 
         // get the query and pass it to the service
-        QueryResultDto queryResult = invoke(
+        AqlQueryRequest aqlQueryRequest = createRequest(
                 queryString,
                 queryParameters,
                 Optional.ofNullable(fetch).map(Integer::longValue),
                 Optional.ofNullable(offset).map(Integer::longValue));
+        QueryResultDto aqlQueryResult = aqlQueryService.query(aqlQueryRequest);
 
         // use the fully qualified metadata location
         Stream<String> pathSegments =
@@ -196,7 +194,7 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
         URI locationUri = createLocationUri(pathSegments.toArray(String[]::new));
 
         // create and return response
-        QueryResponseData queryResponseData = createQueryResponse(queryString, queryResult, locationUri);
+        QueryResponseData queryResponseData = createQueryResponse(aqlQueryResult, queryString, locationUri);
         setQueryName(queryDefinition, queryResponseData);
 
         HttpRestContext.register(QUERY_ID, queryDefinition.getQualifiedName());
@@ -233,10 +231,11 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
         }
 
         // get the query and pass it to the service
-        QueryResultDto queryResult = invoke(queryString, queryRequest);
+        AqlQueryRequest aqlQueryRequest = createRequest(queryString, queryRequest);
+        QueryResultDto aqlQueryResult = aqlQueryService.query(aqlQueryRequest);
 
         // create and return response
-        QueryResponseData queryResponseData = createQueryResponse(queryString, queryResult);
+        QueryResponseData queryResponseData = createQueryResponse(aqlQueryResult, queryString, null);
         setQueryName(queryDefinition, queryResponseData);
 
         HttpRestContext.register(QUERY_ID, queryDefinition.getQualifiedName());
@@ -249,28 +248,45 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
                 QUERY_EXECUTE_ENDPOINT,
                 Boolean.TRUE,
                 HttpRestContext.LOCATION,
-                fromPath("").pathSegment(QUERY, qualifiedName, version).build().toString());
+                fromPath("")
+                        .pathSegment(Q_PARAM, qualifiedName, version)
+                        .build()
+                        .toString());
     }
 
     @SuppressWarnings("unchecked")
-    private QueryResultDto invoke(@NonNull String queryString, Map<String, Object> requestBody) {
+    private AqlQueryRequest createRequest(@NonNull String queryString, Map<String, Object> requestBody) {
 
         requestBody = Optional.ofNullable(requestBody).orElseGet(Map::of);
         Map<String, Object> queryParameters = Optional.ofNullable(requestBody.get(QUERY_PARAMETERS))
                 .map(p -> (Map<String, Object>) p)
                 .orElseGet(Map::of);
-        Optional<Long> fetch = optionalLong(FETCH, requestBody);
-        Optional<Long> offset = optionalLong(OFFSET, requestBody);
+        Optional<Long> fetch = optionalLong(FETCH_PARAM, requestBody);
+        Optional<Long> offset = optionalLong(OFFSET_PARAM, requestBody);
 
-        return invoke(queryString, queryParameters, fetch, offset);
+        return createRequest(queryString, queryParameters, fetch, offset);
     }
 
-    private QueryResultDto invoke(
+    private AqlQueryRequest createRequest(
             @NonNull String queryString, Map<String, Object> parameters, Optional<Long> fetch, Optional<Long> offset) {
 
-        AqlQueryRequest aqlQueryRequest =
-                new AqlQueryRequest(queryString, parameters, fetch.orElse(null), offset.orElse(null));
-        return aqlQueryService.query(aqlQueryRequest);
+        return new AqlQueryRequest(queryString, parameters, fetch.orElse(null), offset.orElse(null));
+    }
+
+    protected QueryResponseData createQueryResponse(
+            QueryResultDto aqlQueryResult, String queryString, @Nullable URI location) {
+        final QueryResponseData queryResponseData = new QueryResponseData(aqlQueryResult);
+        queryResponseData.setQuery(queryString);
+        queryResponseData.setMeta(aqlQueryContext.createMetaData(location));
+        return queryResponseData;
+    }
+
+    // --- Helper ---
+
+    private static void setQueryName(
+            QueryDefinitionResultDto queryDefinitionResultDto, QueryResponseData queryResponseData) {
+        queryResponseData.setName(
+                queryDefinitionResultDto.getQualifiedName() + "/" + queryDefinitionResultDto.getVersion());
     }
 
     private static Optional<Long> optionalLong(String name, Map<String, Object> params) {
@@ -286,38 +302,5 @@ public class OpenehrQueryController extends BaseController implements QueryApiSp
             }
             default -> throw new InvalidApiParameterException("invalid '%s' value '%s'".formatted(name, o));
         });
-    }
-
-    private static void setQueryName(
-            QueryDefinitionResultDto queryDefinitionResultDto, QueryResponseData queryResponseData) {
-        queryResponseData.setName(
-                queryDefinitionResultDto.getQualifiedName() + "/" + queryDefinitionResultDto.getVersion());
-    }
-
-    private QueryResponseData createQueryResponse(String requestedAQL, QueryResultDto resultDto) {
-        return createQueryResponse(requestedAQL, resultDto, null);
-    }
-
-    protected QueryResponseData createQueryResponse(
-            @SuppressWarnings("unused") String requestedAQL, QueryResultDto resultDto, @Nullable URI location) {
-
-        // FIXME adjust org.ehrbase.openehr.sdk:response.dto to not use the query = QueryResultDTO.executedAQL
-        //      because that's wrong - the QueryResponseData.query shall be the original input aql not the performed
-        // one.
-        final QueryResponseData queryResponseData = new QueryResponseData(resultDto);
-        // FIXME this stores the requested AQL query in the response `p` properties - this will be done in an upcoming
-        //       ticket
-        // queryResponseData.setQuery(requestedAQL);
-        MetaData metaData = new MetaData(resultDto);
-        metaData.setHref(Optional.ofNullable(location).map(URI::toASCIIString).orElse(null));
-        // should be extracted from an application property. The current value is 1.0.4
-        metaData.setSchemaVersion(RmConstants.RM_VERSION_1_0_4);
-        queryResponseData.setMeta(metaData);
-
-        if (generatorDetailsEnabled) {
-            metaData.setGenerator("EHRBase/%s".formatted(statusService.getEhrbaseVersion()));
-        }
-
-        return queryResponseData;
     }
 }